requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
boto3==1.38.18
botocore==1.38.18
click==8.2.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-querycount==0.8.3
django-simple-history==3.8.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
